api-book/chapter-7-final-serving/ML_API_docker/requirements.txt
fastapi==0.70.0
uvicorn==0.15.0
gunicorn==20.1.0
xgboost==1.5.0
psycopg2-binary==2.9.2
sqlalchemy-utils==0.37.9
pyaml==21.10.1
cryptography==36.0.1
PyJWT==2.3.0
pandas==1.3.4
scikit-learn==1.0.1